Chassis Sprite Sheets for Animation
So I decided to keep the completed chassis in as a separate asset a) for
reference when piecing the parts together, and b) to use as a completed
sprite for the menu screen, where it doesn't need to animate. The spider
tank is currently in beta, mainly because I want to see what it looks
like while actually moving around before I commit to too much more
detail.